Move node components
