+++ app/src/google/java/com/geeksville/mesh/ui/map/components/NodeClusterMarkers.kt
@@ -21,13 +21,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.key
 import androidx.compose.ui.graphics.Color
 import com.geeksville.mesh.ui.map.NodeClusterItem
-import com.geeksville.mesh.ui.node.components.NodeChip
 import com.google.maps.android.clustering.Cluster
 import com.google.maps.android.clustering.view.DefaultClusterRenderer
 import com.google.maps.android.compose.Circle
 import com.google.maps.android.compose.MapsComposeExperimentalApi
 import com.google.maps.android.compose.clustering.Clustering
 import org.meshtastic.feature.map.BaseMapViewModel
+import org.meshtastic.feature.node.component.NodeChip
 
 @OptIn(MapsComposeExperimentalApi::class)
 @Suppress("NestedBlockDepth")
