+++ app/src/main/java/com/geeksville/mesh/ui/common/components/MainAppBar.kt
@@ -44,12 +44,12 @@ import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.geeksville.mesh.ui.debug.DebugMenuActions
-import com.geeksville.mesh.ui.node.components.NodeChip
 import org.meshtastic.core.database.model.Node
 import org.meshtastic.core.navigation.ContactsRoutes
 import org.meshtastic.core.navigation.SettingsRoutes
 import org.meshtastic.core.strings.R
 import org.meshtastic.core.ui.theme.AppTheme
+import org.meshtastic.feature.node.component.NodeChip
 
 @Suppress("CyclomaticComplexMethod")
 @Composable
