+++ app/src/main/java/com/geeksville/mesh/ui/settings/radio/CleanNodeDatabaseScreen.kt
@@ -46,9 +46,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import com.geeksville.mesh.ui.node.components.NodeChip
 import org.meshtastic.core.database.entity.NodeEntity
 import org.meshtastic.core.strings.R
+import org.meshtastic.feature.node.component.NodeChip
 
 /**
  * Composable screen for cleaning the node database. Allows users to specify criteria for deleting nodes. The list of
