+++ core/ui/src/main/res/drawable/ic_antenna_24.xml
