+++ core/ui/src/main/res/drawable/ic_lock_open_right_24.xml
