+++ feature/node/src/main/kotlin/org/meshtastic/feature/node/component/NodeChip.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.ui.node.components
+package org.meshtastic.feature.node.component
 
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.IntrinsicSize
@@ -75,7 +75,7 @@ fun NodeChip(modifier: Modifier = Modifier, node: Node, onClick: (Node) -> Unit 
 @Suppress("MagicNumber")
 @Preview
 @Composable
-fun NodeChipPreview() {
+private fun NodeChipPreview() {
     val user = MeshProtos.User.newBuilder().setShortName("\uD83E\uDEE0").setLongName("John Doe").build()
     val node =
         Node(
