+++ feature/node/src/main/kotlin/org/meshtastic/feature/node/component/NodeMenu.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.ui.node.components
+package org.meshtastic.feature.node.component
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
